Get rid of ItemsView component, just use ItemsList
The previous iteration smacked of excessive boilerplate. We still
maintain ItemsList as a pure-functional, stateless component. Now we
just wrap it in ItemsContext.Consumer inline in the export. We're
effectively asserting that it will always be used in a state-consuming
way, but keeping its implementation separate from any state-related glue
code.