Add support for images as a separate resource
PNG images can be uploaded via POST at /api/images. An uploaded image
will be encoded as a base64 string and stored as an Image resource in
the database, along with an md5 hash of the original data to prevent
redundant uploads. Images are served (as image/png) from
/api/images/<image_id>.

An Item can have an image associated with it, referenced by ID.

`pipenv run request -t 4 -i foo.png` will upload foo.png and set it as
the image for Item 4.