Add tasks to frontend
In the main (and only) view, listing all the Items in the API, each Item
now lists all the tasks associated with it, via a list of Task
components. Each Task component shows the ID, progress percentage,
status message, and finished/succeeded/failed state of the associated
task, with the help of a ProgressBar component.

The (finger-quotes) "single source of truth" for Tasks lies in a new
React Context object called TasksContext, which is effectively identical
to ItemsContext but with the API endpoint and SocketIO event names
changed.

So at top-level in our app, we have two Providers: one for the
ItemsContext and one for the TasksContext. Then, there are two sets of
Consumers at points where that state needs to be sorted/filtered and
displayed. The ItemsList consumes the state from the ItemsContext, and
each TasksList consumes the state from the TasksContext.

There are multiple TasksList components - one per item - and this
approach means that each one of them consumes the entire state of all
tasks known to the application. Each one simply filters that list of
tasks to only the relevant ones. This seems more sensible than what I've
done previously: it preserves the "single source of truth", and it *just
works* without requiring any special-case frontend logic.

The approach I've taken previously (in a roughly analogous situation)
would have been to associate Tasks with Items via the API. So GETting a
Task would return something like this:

    {
        'id': 4,
	'description': 'This is an item',
	'weight': 0.42,
	'image_id': 1,
	'tasks': [
	    {
	        'id': 1,
		'item_id': 4,
		'started_time': '2019-05-07T23:10:44.969768',
		'finished_time': '2011-05-07T23:11:02.300123',
		'message': 'This task is finished',
		'progress_pct': 1.0,
		'successful': true
	    },
	    {
	        'id': 3,
	        'item_id': 4,
		'started_time': 2019-05-08T15:55:00.501023',
		'finished_time': null,
		'message': 'This task is still running...',
		'progress_pct': 0.2333,
		'successful': null
	    }
	]
    }

Contrast this old approach with the one employed in this change, where
we simply provide a flat list of items via GET /api/items, and a flat
list of  tasks via GET /api/tasks, and we leave it up to the frontend to
associate them properly.

The old approach simplifies the API on the surface, and perhaps it has
some merits in cases where we want to optimize for the lowest possible
number of HTTP requests. However, the dependencies that it creates lead
to some messy complexity:

On the backend, when a task is updated, created, or deleted, then we
have to traverse the dependency graph to figure out which items are
affected by that change, and notify the frontend that they've changed
and need to be updated. Since emitting 'item update' means we
re-serialize and transfer the entire item, this approach is far less
bandwidth-efficient in practice.