Implement POST /api/images/<id>, cli replace-image
This is a somewhat contrived bit of functionality: since we enforce the
invariant that all uploaded images are unique (i.e. if a client uploads
the same image twice, the first request will create and store the image,
and the second will just spit back the same image resource), the image
hash (and PNG data) are effectively immutable once stored. So it
wouldn't make sense to allow users to directly modify the data
associated with an image resource via POST.

Instead, we support POST on an individual image, which is implemented as
a slightly more complex operation under the hood, whereby the relevant
items have their image references updated.

This simply gives us more surface area to test with, so we can simulate
some of the real-world complexities of an actual app.