Add Event resource with API
Events can be created via POST to /api/events. An event has a string
message and an optional severity level (defaults to info). Events are
timestamped on creation and are not meant to be edited after creation.

Events may also be created by the app itself, in response to other API
calls or internal operations.

The introduction of Events gives us more complexity to test with. For
one, we can now introduce a separate view from the items list. Just as
importantly, events are meant to be numerous enough that we can't expect
to keep track of all of them: GET requests to /api/events yield a result
that's paginated, so our frontend code will now have to deal with
paginated results (for events) as well as comprehensive, flat lists
(for items and images).