Add WIP events support to frontend
This change adds an EventsContext and an EventsList component that
renders a series of Event components for all events. Note that
EventsProvider is currently incomplete, since it doesn't handle
paginated results: it simply displays the first page of results as if
it's the complete set of events.

Switching between the Items view and the Events view is not supported
at the moment either: there's simply a hardcoded toggle in index.jsx.